+++ src/components/CancelledRouteCube.tsx
@@ -1,6 +1,4 @@
 import styled from "styled-components";
-import { Stop } from './types';
-
 
 const StyledCube = styled.div`
   height: 50px;
@@ -15,24 +13,23 @@ const StyledCube = styled.div`
 `;
 
 const WrapperWithDiagonalLine = styled.div`
-  width: calc(80px*1.41);
+  width: calc(80px * 1.41);
   border-bottom: 3px solid red;
   transform: rotate(135deg);
   transform-origin: left;
 `;
 
-
-interface CancalledRouteCubeProps {
+interface ICancelledRouteCubeProps {
   text: string;
   color: string;
 }
 
-const CancalledRouteCubeProps = ({ text, color}: CancalledRouteCubeProps) => {
+const CancelledRouteCubeProps = ({ text, color }: ICancelledRouteCubeProps) => {
   return (
-   <WrapperWithDiagonalLine>
-    <StyledCube color={color}> {text} </StyledCube>
-   </WrapperWithDiagonalLine>
+    <WrapperWithDiagonalLine>
+      <StyledCube color={color}> {text} </StyledCube>
+    </WrapperWithDiagonalLine>
   );
 };
 
-export default CancalledRouteCubeProps;
+export default CancelledRouteCubeProps;
